requirements.txt
Flask==3.0.0
requests
python-dotenv
Werkzeug==3.0.0
gunicorn
